backend/requirements.txt
langgraph
pydantic-ai
langchain-openai
python-dotenv
tabulate
fastapi
uvicorn